REF/TST: collect misplaced tests
