+++ pandas/tests/indexes/multi/test_integrity.py
@@ -250,25 +250,6 @@ def test_rangeindex_fallback_coercion_bug():
     tm.assert_index_equal(result, expected)
 
 
-def test_hash_error(indices):
-    index = indices
-    with pytest.raises(TypeError, match=f"unhashable type: '{type(index).__name__}'"):
-        hash(indices)
-
-
-def test_mutability(indices):
-    if not len(indices):
-        return
-    msg = "Index does not support mutable operations"
-    with pytest.raises(TypeError, match=msg):
-        indices[0] = indices[0]
-
-
-def test_wrong_number_names(indices):
-    with pytest.raises(ValueError, match="^Length"):
-        indices.names = ["apple", "banana", "carrot"]
-
-
 def test_memory_usage(idx):
     result = idx.memory_usage()
     if len(idx):
